feat(docker): add bc package to build image
Added the 'bc' package to the build Dockerfile to support scripts or
tools
that require basic calculator functionality in the CI/CD environment.
Also updated Makefile to use a versioned image tag and renamed the
build target to 'publish'. Removed the unused Dockerfile.mac and moved
the start-mac.sh script to the build directory.